refactor(endpoints): rename `kw` mod to `keyword_search`
