+++ crates/endpoints/src/lib.rs
@@ -15,7 +15,7 @@ pub mod files;
 pub mod images;
 #[cfg(feature = "index")]
 #[cfg_attr(docsrs, doc(cfg(feature = "index")))]
-pub mod kw;
+pub mod keyword_search;
 pub mod models;
 #[cfg(feature = "rag")]
 #[cfg_attr(docsrs, doc(cfg(feature = "rag")))]
